Javascript functionality added in delete button.

+++ admin/includes/view_all_comments.php
@@ -93,7 +93,7 @@ $select_comments = mysqli_query($connection, $query);
 
                        
 
-                        echo "<td><a href ='comments.php?delete=$comment_id'>Delete</a></td>";
+                        echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete'); \" href ='comments.php?delete=$comment_id'>Delete</a></td>";
 
                          echo "</tr>";
